// src/models/LFancy.ts
export default interface LFancy {
  _id?: string
  active: boolean
  fancyName: string
  marketId: string
  matchId: number
  sportId: string
  GameStatus?: string
  gtype: string
  sr_no: number
  isSuspend?: boolean
  createdAt?: string
  updatedAt?: string
  result: string
  bet?: boolean
  rem?:string
}